requirements.txt
python-binance
python-telegram-bot==20.7
python-dotenv
aiohttp
websockets==10.4
